refactor: moved some event classes
