+++ src/main/java/zone/vao/nexoAddon/NexoAddon.java
@@ -37,6 +37,9 @@ import zone.vao.nexoAddon.events.*;
 import zone.vao.nexoAddon.events.blocks.BlockBreakListener;
 import zone.vao.nexoAddon.events.blocks.BlockPlaceListener;
 import zone.vao.nexoAddon.events.blocks.NexoStringBlockInteractListener;
+import zone.vao.nexoAddon.events.inventoryClicks.InventoryClickListener;
+import zone.vao.nexoAddon.events.nexo.NexoItemsLoadedListener;
+import zone.vao.nexoAddon.events.nexo.NexoPackUploadListener;
 import zone.vao.nexoAddon.events.nexo.blocks.NexoBlockBreakListener;
 import zone.vao.nexoAddon.events.nexo.blocks.NexoBlockInteractListener;
 import zone.vao.nexoAddon.events.nexo.blocks.NexoBlockPlaceListener;
