+++ src/main/java/zone/vao/nexoAddon/events/nexo/NexoItemsLoadedListener.java
@@ -1,4 +1,4 @@
-package zone.vao.nexoAddon.events;
+package zone.vao.nexoAddon.events.nexo;
 
 import com.nexomc.nexo.api.events.NexoItemsLoadedEvent;
 import lombok.Getter;
